requirements.txt
keras==2.9.0
matplotlib==3.5.2
numpy==1.23.0
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.2.0
streamlit==1.10.0
